Requested changes by @Serabe
